Make it idiomatic
